Moving tests per recommendation.
